refactor: Переименование компоненты, рефакторинг сборки
