+++ src/OneScriptCompressor/BrotliCompressor.cs
@@ -10,7 +10,7 @@ using BrotliSharpLib;
 using OneScript.Contexts;
 #endif
 
-namespace OscriptCompressor
+namespace OneScriptCompressor
 {
     /// <summary>
     /// Класс для упаковки/распаковки данных в Brotli.
