+++ src/OneScriptCompressor/GZipCompressor.cs
@@ -6,7 +6,7 @@ using ScriptEngine.Machine.Contexts;
 using OneScript.Contexts;
 #endif
 
-namespace OscriptCompressor
+namespace OneScriptCompressor
 {
     /// <summary>
     /// Класс для упаковки/распаковки данных в GZip.
