+++ src/OneScriptCompressor/LZ4Compressor.cs
@@ -7,7 +7,7 @@ using EasyCompressor;
 using OneScript.Contexts;
 #endif
 
-namespace OscriptCompressor
+namespace OneScriptCompressor
 {
     /// <summary>
     /// Класс для упаковки/распаковки данных в LZ4.
