+++ src/OneScriptCompressor/OneScriptCompressor.csproj
@@ -5,7 +5,7 @@
 		<LangVersion>10.0</LangVersion>
 		<PackageProjectUrl>https://github.com/Stivo182/oscript-compressor</PackageProjectUrl>
 		<RepositoryUrl>https://github.com/Stivo182/oscript-compressor</RepositoryUrl>
-		<Title>OscriptCompressor</Title>
+		<Title>OneScriptCompressor</Title>
 		<Company>Stivo182</Company>
 		<Copyright>Stivo182</Copyright>
 		<Version>1.0.1</Version>
