+++ src/OneScriptCompressor/SnappyCompressor.cs
@@ -6,7 +6,7 @@ using System.IO;
 using OneScript.Contexts;
 #endif
 
-namespace OscriptCompressor
+namespace OneScriptCompressor
 {
     /// <summary>
     /// Класс для упаковки/распаковки данных в Snappy.
