+++ src/OneScriptCompressor/ZstdCompressionLevel.cs
@@ -4,7 +4,7 @@ using ScriptEngine;
 using OneScript.Contexts.Enums;
 #endif
 
-namespace OscriptCompressor
+namespace OneScriptCompressor
 {
     /// <summary>
     /// ZStandard Compression Level
