+++ src/OneScriptCompressor/ZstdCompressor.cs
@@ -5,7 +5,7 @@ using ScriptEngine.Machine.Contexts;
 using OneScript.Contexts;
 #endif
 
-namespace OscriptCompressor
+namespace OneScriptCompressor
 {
     /// <summary>
     /// Класс для упаковки/распаковки данных в ZStandard.
