+++ tasks/oscript.cfg
@@ -1,0 +1,1 @@
+lib.system=../oscript_modules
