+++ tests/compressor_test.os
@@ -1,5 +1,5 @@
 #Использовать asserts
-#Использовать compressor
+#Использовать ".."
 
 Перем МенеджерВременныхФайлов;
 Перем Потоки;
